renamed edit_profile use case to EditProfile for consistency
