+++ src/main/java/app/EditUseCaseFactory.java
@@ -6,10 +6,10 @@ import interface_adapter.edit_profile.EditController;
 import interface_adapter.edit_profile.EditPresenter;
 import interface_adapter.edit_profile.EditViewModel;
 import interface_adapter.edited_profile.EditedViewModel;
-import use_case.edit_profile.EditInputBoundary;
-import use_case.edit_profile.EditInteractor;
-import use_case.edit_profile.EditOutputBoundary;
-import use_case.edit_profile.EditPatientDataAccessInterface;
+import use_case.EditProfile.EditInputBoundary;
+import use_case.EditProfile.EditInteractor;
+import use_case.EditProfile.EditOutputBoundary;
+import use_case.EditProfile.EditPatientDataAccessInterface;
 import view.EditView;
 
 import javax.swing.*;
