+++ src/test/java/use_case/EditProfile/EditProfileInteractorTest.java
@@ -2,7 +2,6 @@ package use_case.EditProfile;
 
 import data_access.FilePatientDataAccessObject;
 import entities.*;
-import use_case.edit_profile.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
